+++ react/helpers/getParsedAddress.tsx
@@ -32,6 +32,7 @@ export const getParsedAddress = (place: { address_components: any[] }) => {
     addressType: 'residential',
     city:
       parsedAddressComponents.administrative_area_level_3 ||
+      parsedAddressComponents.administrative_area_level_2 ||
       parsedAddressComponents.locality,
     // complement: '',
     /* Google Maps API returns Alpha-2 ISO codes, but checkout API requires Alpha-3 */